requirements.txt
flask
flask_pymongo
graphene
beautifulsoup4
requests
flask_graphql
PyJWT
